Move miniz files to 3rdparty/miniz to avoid clang-tidy checks
